+++ CHANGELOG.md
@@ -50,6 +50,7 @@ With #144 you can now join streams with differing group by dimensions.
 - [#215](https://github.com/influxdata/kapacitor/issues/215): Add performance metrics to nodes for average execution times and node throughput values.
 - [#144](https://github.com/influxdata/kapacitor/issues/144): Can now join streams with differing dimensions using the join.On property.
 - [#249](https://github.com/influxdata/kapacitor/issues/249): Can now use InfluxQL functions directly instead of via the MapReduce method. Example `stream.from().count()`.
+- [#302](https://github.com/influxdata/kapacitor/issues/302): Can now use .Time in alert message.
 
 
 ### Bugfixes